requirements.txt
aiohttp==3.5.4
asyncpg==0.18.3
python-dateutil==2.8.0
numpy==1.17.0
pytest-asyncio==0.10.0